Name the INI ModifierList block parser
The exe's block registry roots the keyword ModifierList at 0x0036B7C0, which
the ledger carried as parseRva00369860State in Rva0036B7C0ParseState.cpp --
nothing a modder editing a Leadership bonus could find. Rename it to
parseModifierListDefinition and move the file next to the other block parsers
as Common/INI/INIModifierList.cpp.

The definition class it allocates is the one already matched as
AttributeModifierDefinition in Rva00369860Reset.cpp, but its ctor and STLport
insertion pins in reverse/symbols.csv still spell it Rva00369860State, so
renaming the class is a separate change across both translation units.

Also lands two sibling lanes' finished file renames (0x00677530 and 0x002BEE90)
that were sitting in the shared tree. Their ledger rows already pointed at the
new paths, and functions.csv is one file: nobody could commit it while a row
named a source that was not yet in HEAD, so this breaks that standoff. Both
files are byte-verified by this commit's gate.
